Add DataService.py, update __main__.py
